Tweaked scripts

+++ scripts/apps/BridgeArduinoSerialToMqttSplitCsv/init.sh
@@ -18,10 +18,10 @@ if [ -f $CONFIG_FILE ]; then
 fi
 
 # TODO: Clean up. This check is disabled to allow the install package script to be overwritten
-#if [ ! -f "install-package.sh" ]; then
+if [ ! -f "install-package.sh" ]; then
   INSTALL_SCRIPT_FILE_URL="https://raw.githubusercontent.com/GreenSense/Index/master/scripts/apps/BridgeArduinoSerialToMqttSplitCsv/install-package.sh"
   wget --no-cache -O install-package.sh $INSTALL_SCRIPT_FILE_URL
-#fi
+fi
 
 sh install-package.sh BridgeArduinoSerialToMqttSplitCsv 1.0.1.3 || exit 1
